# openapi.yaml
openapi: 3.0.4
info:
    title: Weather API
    description: Returns all meteorological stations or a single station's info
    version: "1.0"

servers:
    - url: http://localhost:8000/api
      description: Local server

paths:
    /stations:
        get:
            summary: Returns a list of stations with name and station_id.
            responses:
                "200": # status code
                    description: A JSON array of stations
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        STATION_ID:
                                            type: string
                                        NAME:
                                            type: string
                "404": # status code
                    description: A JSON error message if no stations are returned
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                "500": # status code
                    description: A JSON error message if an internal error occurred
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
    /stations/{stationId}:
        get:
            summary: Returns a particular station's info.
            parameters:
                -   name: stationId
                    in: path
                    required: true
                    schema:
                        type: string
                    description: Station identifier
            responses:
                "200": # status code
                    description: A JSON array of one station's fields
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        _id:
                                            type: integer
                                        STATION_ID:
                                            type: string
                                        NAME:
                                            type: string
                                        WMO_ID:
                                            type: string
                                        BEGIN_DATE:
                                            type: string
                                            format: date
                                        END_DATE:
                                            type: string
                                            format: date
                                        LATITUDE:
                                            type: integer
                                        LONGITUDE:
                                            type: integer
                                        GAUSS1:
                                            type: string
                                        GAUSS2:
                                            type: string
                                        GEOGR1:
                                            type: number
                                        GEOGR2:
                                            type: number
                                        ELEVATION:
                                            type: number
                                        ELEVATION_PRESSURE:
                                            type: string
                                        "rank STATION_ID":
                                            type: number
                "400": # status code
                    description: A JSON error message if stationId is invalid (longer than 4000 characters or contains whitespace)
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                "404": # status code
                    description: A JSON error message if no station is found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                "500": # status code
                    description: A JSON error message if an internal error occurred
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string